+++ CHANGES.md
@@ -5,6 +5,8 @@
 * @dev (????-??-??)
   * Added Telnet component
   * Fixed TTY appearance bug while changing various zoom levels.
+  * Added Metal graphics acceleration
+  * Added option to hide/show toolbar
 
 * v3.9.0 (2024-08-15)
   * Updated Java requirement to Java 21.